Fixes #143644: Use `strict` in `TestInstantiationService`
